+++ navground_sim/include/navground/sim/state_estimations/sensor_combination.h
@@ -59,7 +59,7 @@ struct NAVGROUND_SIM_EXPORT SensorCombination : public Sensor {
    * @private
    */
   virtual void update(Agent *agent, World *world,
-                      EnvironmentState *state) const override;
+                      EnvironmentState *state) override;
 
   Description get_description() const override;