+++ navground_sim/include/navground/sim/state_estimations/sensor_discs.h
@@ -9,9 +9,9 @@
 #include <vector>
 
 #include "navground/core/types.h"
+#include "navground/sim/export.h"
 #include "navground/sim/state_estimations/sensor.h"
 #include "navground/sim/world.h"
-#include "navground/sim/export.h"
 
 using navground::core::BufferDescription;
 using navground::core::make_property;
@@ -31,13 +31,15 @@ namespace navground::sim {
  *
  *   - `number` (int, \ref get_number)
  *
- *   - `max_radius` (int, \ref get_number)
+ *   - `max_radius` (int, \ref get_max_radius)
  *
- *   - `max_speed` (int, \ref get_number)
+ *   - `max_speed` (int, \ref get_max_speed)
+ *
+ *   - `include_valid` (bool, \ref get_include_valid)
  *
  *   - `use_nearest_point` (bool, \ref get_use_nearest_point)
  *
- *   - `max_
+ *   - `max_id` (int, \ref get_max_id)
  *
  */
 struct NAVGROUND_SIM_EXPORT DiscsStateEstimation : public Sensor {
@@ -50,11 +52,11 @@ struct NAVGROUND_SIM_EXPORT DiscsStateEstimation : public Sensor {
    */
   inline static const ng_float_t default_number = 1;
   /**
-   * The default maximal neighbor radius (for which it will not include radii)
+   * The default maximal neighbor radius (zero means that it will include radii)
    */
   inline static const ng_float_t default_max_radius = 0;
   /**
-   * The default maximal neighbor speed (for which it will not include
+   * The default maximal neighbor speed (zero means that it will not include
    * velocities)
    */
   inline static const ng_float_t default_max_speed = 0;
@@ -67,30 +69,31 @@ struct NAVGROUND_SIM_EXPORT DiscsStateEstimation : public Sensor {
    */
   inline static const bool default_use_nearest_point = true;
   /**
-   * The default maximal id (for which it will not include ids)
+   * The default maximal id (zero means that it will not include ids)
    */
   inline static const unsigned default_max_id = 0;
   /**
    * @brief      Constructs a new instance.
    *
-   * @param[in]  range_   The range of view
-   * @param[in]  number_  Number of discs
+   * @param[in]  range   The range of view
+   * @param[in]  number  Number of discs
+   * @param[in]  max_radius  Maximal neighbor radius
+   * @param[in]  max_speed  Maximal neighbor speed
+   * @param[in]  include_valid  Whether to include validity
+   * @param[in]  use_nearest_point  Whether to use nearest point as position
+   * @param[in]  max_id  Maximal neighbor id
+   * @param[in]  name     The name to use as a prefix
    */
   explicit DiscsStateEstimation(
-      ng_float_t range_ = default_range, unsigned number_ = default_number,
-      ng_float_t max_radius_ = default_max_radius,
-      ng_float_t max_speed_ = default_max_speed,
-      bool include_valid_ = default_include_valid,
-      bool use_nearest_point_ = default_use_nearest_point,
-      unsigned max_id_ = default_max_id)
-      : Sensor(),
-        range(range_),
-        number(number_),
-        max_radius(max_radius_),
-        max_speed(max_speed_),
-        include_valid(include_valid_),
-        use_nearest_point(use_nearest_point_),
-        max_id(max_id_) {}
+      ng_float_t range = default_range, unsigned number = default_number,
+      ng_float_t max_radius = default_max_radius,
+      ng_float_t max_speed = default_max_speed,
+      bool include_valid = default_include_valid,
+      bool use_nearest_point = default_use_nearest_point,
+      unsigned max_id = default_max_id, const std::string &name = "")
+      : Sensor(name), _range(range), _number(number), _max_radius(max_radius),
+        _max_speed(max_speed), _include_valid(include_valid),
+        _use_nearest_point(use_nearest_point), _max_id(max_id) {}
 
   virtual ~DiscsStateEstimation() = default;
 
@@ -99,28 +102,28 @@ struct NAVGROUND_SIM_EXPORT DiscsStateEstimation : public Sensor {
    *
    * @param[in]  value     The new value
    */
-  void set_range(ng_float_t value) { range = std::max<ng_float_t>(0, value); }
+  void set_range(ng_float_t value) { _range = std::max<ng_float_t>(0, value); }
 
   /**
    * @brief      Gets the range of view.
    *
    * @return     The range of view.
    */
-  ng_float_t get_range() const { return range; }
+  ng_float_t get_range() const { return _range; }
 
   /**
    * @brief      Sets the number of discs.
    *
    * @param[in]  value     The new value
    */
-  void set_number(int value) { number = std::max(0, value); }
+  void set_number(int value) { _number = std::max(0, value); }
 
   /**
    * @brief      Gets the number of discs.
    *
    * @return     The number of discs.
    */
-  int get_number() const { return number; }
+  int get_number() const { return _number; }
 
   /**
    * @brief      Sets the maximal neighbor radius.
@@ -128,7 +131,7 @@ struct NAVGROUND_SIM_EXPORT DiscsStateEstimation : public Sensor {
    * @param[in]  value     The new value
    */
   void set_max_radius(ng_float_t value) {
-    max_radius = std::max<ng_float_t>(0, value);
+    _max_radius = std::max<ng_float_t>(0, value);
   }
 
   /**
@@ -136,7 +139,7 @@ struct NAVGROUND_SIM_EXPORT DiscsStateEstimation : public Sensor {
    *
    * @return     The maximal radius.
    */
-  ng_float_t get_max_radius() const { return max_radius; }
+  ng_float_t get_max_radius() const { return _max_radius; }
 
   /**
    * @brief      Sets the maximal neighbor speed.
@@ -144,7 +147,7 @@ struct NAVGROUND_SIM_EXPORT DiscsStateEstimation : public Sensor {
    * @param[in]  value     The new value
    */
   void set_max_speed(ng_float_t value) {
-    max_speed = std::max<ng_float_t>(0, value);
+    _max_speed = std::max<ng_float_t>(0, value);
   }
 
   /**
@@ -152,34 +155,34 @@ struct NAVGROUND_SIM_EXPORT DiscsStateEstimation : public Sensor {
    *
    * @return     The maximal speed.
    */
-  ng_float_t get_max_speed() const { return max_speed; }
+  ng_float_t get_max_speed() const { return _max_speed; }
 
   /**
    * @brief      Sets whether to include validity.
    *
    * @param[in]  value     The new value
    */
-  void set_include_valid(bool value) { include_valid = value; }
+  void set_include_valid(bool value) { _include_valid = value; }
 
   /**
    * @brief      Gets whether to include validity.
    *
    * @return     Whether to include validity.
    */
-  bool get_include_valid() const { return include_valid; }
+  bool get_include_valid() const { return _include_valid; }
   /**
    * @brief      Sets whether to use the nearest or the center as position
    *
    * @param[in]  value     The new value
    */
-  void set_use_nearest_point(bool value) { use_nearest_point = value; }
+  void set_use_nearest_point(bool value) { _use_nearest_point = value; }
 
   /**
    * @brief      Gets whether to use the nearest or the center as position
    *
    * @return     Whether to use nearest point as position.
    */
-  bool get_use_nearest_point() const { return use_nearest_point; }
+  bool get_use_nearest_point() const { return _use_nearest_point; }
 
   /**
    * @brief      Sets the maximal possible id.
@@ -188,14 +191,14 @@ struct NAVGROUND_SIM_EXPORT DiscsStateEstimation : public Sensor {
    *
    * @param[in]  value  The new value
    */
-  void set_max_id(int value) { max_id = std::max(0, value); }
+  void set_max_id(int value) { _max_id = std::max(0, value); }
 
   /**
    * @brief      Gets the maximal possible id.
    *
    * @return     The maximal id.
    */
-  int get_max_id() const { return max_id; }
+  int get_max_id() const { return _max_id; }
 
   /**
    * @private
@@ -218,52 +221,61 @@ struct NAVGROUND_SIM_EXPORT DiscsStateEstimation : public Sensor {
    * @private
    */
   virtual void update(Agent *agent, World *world,
-                      EnvironmentState *state) const override;
+                      EnvironmentState *state) override;
 
   // {(relative x, relative y, v_x, v_y, radius)}
   // DONE(Jerome): max_radius / max_velocity
+
+  /**
+   * @private
+   */
   Description get_description() const override {
     Description desc;
-    if (!number) return desc;
+    if (!_number)
+      return desc;
     if (include_radius()) {
-      desc.emplace("radius", BufferDescription::make<ng_float_t>({number}, 0.0,
-                                                                 max_radius));
+      desc.emplace(
+          get_field_name("radius"),
+          BufferDescription::make<ng_float_t>({_number}, 0, _max_radius));
     }
     if (include_velocity()) {
-      desc.emplace("velocity", BufferDescription::make<ng_float_t>(
-                                   {number, 2}, -max_speed, max_speed));
+      desc.emplace(get_field_name("velocity"),
+                   BufferDescription::make<ng_float_t>(
+                       {_number, 2}, -_max_speed, _max_speed));
     }
     if (include_position()) {
-      desc.emplace("position", BufferDescription::make<ng_float_t>(
-                                   {number, 2}, -range, range));
+      desc.emplace(
+          get_field_name("position"),
+          BufferDescription::make<ng_float_t>({_number, 2}, -_range, _range));
     }
     if (get_include_valid()) {
-      desc.emplace("valid", BufferDescription::make<uint8_t>({number}, 0, 1));
+      desc.emplace(get_field_name("valid"),
+                   BufferDescription::make<uint8_t>({_number}, 0, 1));
     }
     if (include_id()) {
-      desc.emplace(
-          "id", BufferDescription::make<unsigned>({number}, 0, max_id, true));
+      desc.emplace(get_field_name("id"), BufferDescription::make<unsigned>(
+                                             {_number}, 0, _max_id, true));
     }
     return desc;
   }
 
- private:
-  ng_float_t range;
-  unsigned number;
-  ng_float_t max_radius;
-  ng_float_t max_speed;
-  bool include_valid;
-  bool use_nearest_point;
-  unsigned max_id;
+private:
+  ng_float_t _range;
+  unsigned _number;
+  ng_float_t _max_radius;
+  ng_float_t _max_speed;
+  bool _include_valid;
+  bool _use_nearest_point;
+  unsigned _max_id;
   const static std::string type;
 
-  bool include_velocity() const { return max_speed > 0; }
-  bool include_radius() const { return max_radius > 0; }
-  bool include_position() const { return range > 0; }
-  bool include_id() const { return max_id > 0; }
+  bool include_velocity() const { return _max_speed > 0; }
+  bool include_radius() const { return _max_radius > 0; }
+  bool include_position() const { return _range > 0; }
+  bool include_id() const { return _max_id > 0; }
 };
 
-}  // namespace navground::sim
+} // namespace navground::sim
 
-#endif /* end of include guard: \
+#endif /* end of include guard:                                                \
           NAVGROUND_SIM_STATE_ESTIMATIONS_SENSOR_DISCS_H_ */