+++ navground_sim/src/state_estimation.cpp
@@ -8,7 +8,7 @@
 
 namespace navground::sim {
 
-void StateEstimation::update(Agent *agent, World *world) const {
+void StateEstimation::update(Agent *agent, World *world) {
   if (agent) {
     if (Behavior *behavior = agent->get_behavior()) {
       update(agent, world, behavior->get_environment_state());