+++ navground_sim/src/state_estimations/sensor_lidar.cpp
@@ -10,24 +10,61 @@
 namespace navground::sim {
 
 void LidarStateEstimation::update(Agent *agent, World *world,
-                                  EnvironmentState *state) const {
+                                  EnvironmentState *state) {
   if (core::SensingState *_state = dynamic_cast<core::SensingState *>(state)) {
-    auto &c = const_cast<core::CollisionComputation &>(cc);
-    const auto neighbors = world->get_neighbors(agent, range);
-    c.setup(agent->pose, 0.0, world->get_line_obstacles(), world->get_discs(),
+    auto &c = const_cast<core::CollisionComputation &>(_cc);
+    const auto neighbors = world->get_neighbors(agent, _range);
+    const auto pose = core::Pose2(_position, 0).absolute(agent->pose);
+    c.setup(pose, 0.0, world->get_line_obstacles(), world->get_discs(),
             neighbors);
-    const auto ranges = c.get_free_distance_for_sector(
-        agent->pose.orientation + start_angle, field_of_view, resolution - 1,
-        range, false);
-    // _state->set<float>(field_name, ranges, true);
-    auto buffer = _state->get_buffer(field_name);
-    if (!buffer) {
-      buffer =
-          _state->init_buffer(field_name, get_description().at(field_name));
+    auto ranges = c.get_free_distance_for_sector(
+        agent->pose.orientation + _start_angle, _field_of_view, _resolution - 1,
+        _range, false);
+    auto buffer = get_or_init_buffer(*_state, field_name);
+    if (buffer) {
+      if (has_error()) {
+        auto & rg = world->get_random_generator();
+        for (size_t i = 0; i < ranges.size(); ++i) {
+          ranges[i] = std::clamp<ng_float_t>(ranges[i] + _error(rg), 0, _range);
+        }
+      }
+      buffer->set_data(ranges);
     }
-    if (!buffer) return;
-    buffer->set_data(ranges);
   }
+}
+
+// std::valarray<ng_float_t> LidarStateEstimation::sample_error(World *world) {
+//   auto rg = world->get_random_generator();
+//   std::valarray<ng_float_t> vs(_resolution);
+//   for (size_t i = 0; i < _resolution; ++i) {
+//     vs[i] = _error(rg);
+//   }
+//   return vs;
+// }
+
+const std::valarray<ng_float_t> &
+LidarStateEstimation::read_ranges(core::SensingState &state) const {
+  const auto buffer = get_or_init_buffer(state, field_name);
+  return *(buffer->get_data<ng_float_t>());
+}
+
+ng_float_t LidarStateEstimation::get_angular_increment() const {
+  if (_resolution > 1) {
+    return _field_of_view / (_resolution - 1);
+  }
+  return 0;
+}
+
+std::valarray<ng_float_t> LidarStateEstimation::get_angles() const {
+  std::valarray<ng_float_t> vs(_resolution);
+  ng_float_t angle = _start_angle;
+  const ng_float_t delta = get_angular_increment();
+  for (size_t i = 0; i < vs.size() - 1; ++i) {
+    vs[i] = angle;
+    angle += delta;
+  }
+  vs[vs.size() - 1] = _start_angle + _field_of_view;
+  return vs;
 }
 
 const std::map<std::string, Property> LidarStateEstimation::properties =
@@ -48,10 +85,23 @@ const std::map<std::string, Property> LidarStateEstimation::properties =
                            &LidarStateEstimation::get_resolution,
                            &LidarStateEstimation::set_resolution,
                            default_resolution, "Resolution")},
+        {"position", make_property<core::Vector2, LidarStateEstimation>(
+                         &LidarStateEstimation::get_position,
+                         &LidarStateEstimation::set_position,
+                         core::Vector2::Zero(), "Relative position")},
+        {"error_bias", make_property<ng_float_t, LidarStateEstimation>(
+                           &LidarStateEstimation::get_error_bias,
+                           &LidarStateEstimation::set_error_bias,
+                           default_error_bias, "Error bias")},
+        {"error_std_dev",
+         make_property<ng_float_t, LidarStateEstimation>(
+             &LidarStateEstimation::get_error_std_dev,
+             &LidarStateEstimation::set_error_std_dev, default_error_std_dev,
+             "Error standard deviation")},
     } +
-    StateEstimation::properties;
+    Sensor::properties;
 
 const std::string LidarStateEstimation::type =
     register_type<LidarStateEstimation>("Lidar");
 
-}  // namespace navground::sim
+} // namespace navground::sim